ES-121 fix FakeLoader
